requirements.txt
ask-sdk==1.19.0
ask-sdk-core==1.19.0
ask-sdk-dynamodb-persistence-adapter==1.19.0
ask-sdk-model==1.82.0
ask-sdk-runtime==1.19.0
ask-sdk-webservice-support==1.3.3
asn1crypto==1.5.1
boto3==1.34.147
botocore==1.34.147
bottle==0.12.25
certifi==2024.7.4
certvalidator==0.11.1
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
freezegun==1.5.1
gevent==24.2.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
load-dotenv==0.1.0
MarkupSafe==2.1.5
oscrypto==1.3.0
packaging==24.1
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
s3transfer==0.10.2
setuptools==71.1.0
six==1.16.0
urllib3==2.2.2
zope.event==5.0
zope.interface==6.4.post2
